[FIX] syspath loading
Before this commit, the addons path order was not guaranteed to comply
with dockery-odoo folder sorting.

Reasons: odoo.modules.module.ad_paths is initialized at __init__ time,
not only through a server start.

This commit Override it in the decisive moment to reset the addons path.

Additionally, we detected that the config default loading would not take
into account env variable settings for RC files